Refactor DMS conversion to UIUtils
* Add AndroidJUnitRunner unit test for DMS conversion method that requires Android Context
* Tweak strings for Settings text
