+++ GPSTest/build.gradle
@@ -13,6 +13,8 @@ android {
         versionName "3.1.5"
 
         vectorDrawables.useSupportLibrary = true
+
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
 
     flavorDimensions "map"
@@ -82,6 +84,7 @@ android {
 
     testOptions {
         unitTests.returnDefaultValues = true
+        unitTests.includeAndroidResources true
     }
 }
 
@@ -144,4 +147,5 @@ dependencies {
 
     // Unit tests
     testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'com.android.support.test:runner:1.0.2'
 }
